Module common_telemetry::logging

logging stuffs, inspired by databend

Modules

event

Events represent single points in time during the execution of a program.

span

Spans represent periods of time in which a program was executing in a particular context.

Macros

debug

Logs a message at the debug level.

error

Logs a message at the error level.

event

Constructs a new Event.

info

Logs a message at the info level.

log

The standard logging macro.

span

Constructs a new span.

trace

Logs a message at the trace level.

warn

Logs a message at the warn level.

Structs

Level

Describes the level of verbosity of a span or event.

LoggingOptions

TracingOptions

Constants

DEFAULT_LOG_TARGETS 🔒

Statics

GLOBAL_UT_LOG_GUARD 🔒

TRACE_ID

Task local trace id. See trace_id for more details.

Functions

init_default_ut_logging

Init tracing for unittest. Write logs to file unittest.

init_global_logging

trace_id

Get current TRACE_ID from tokio task_local storage.