common_telemetry

Module logging

logging stuffs, inspired by databend

Structs

LoggingOptions

The logging options that used to initialize the logger.

SlowQueryOptions

The options of slow query.

TraceContext 🔒

TraceLayer 🔒

A tracing layer that can be dynamically reloaded.

TraceReloadHandle

TracingOptions

Enums

LogFormat

OtlpExportProtocol

The protocol of OTLP export.

SlowQueriesRecordType

TraceState 🔒

Constants

DEFAULT_LOGGING_DIR

The default logs directory.

DEFAULT_LOG_TARGETS 🔒

DEFAULT_OTLP_GRPC_ENDPOINT

The default endpoint when use gRPC exporter protocol.

DEFAULT_OTLP_HTTP_ENDPOINT

The default endpoint when use HTTP exporter protocol.

Statics

GLOBAL_UT_LOG_GUARD 🔒

LOG_RELOAD_HANDLE

Handle for reloading log level

TRACER 🔒

TRACE_RELOAD_HANDLE

Handle for reloading trace level

Functions

build_otlp_exporter 🔒

build_slow_query_logger 🔒

create_tracer 🔒

get_or_init_tracer

Ensure that the OTLP tracer has been constructed, building it lazily if needed.

init_default_ut_logging

Init tracing for unittest. Write logs to file unittest.

init_global_logging

Type Aliases

DynSubscriber 🔒

OtelTraceLayer 🔒