common_telemetry

Module tracing_sampler

Structs

TracingSampleOptions

TracingSampleRule

Determine the sampling rate of a span according to the rules provided in RuleSampler. For spans that do not hit any rules, the default_ratio is used.

Functions

create_sampler

sample_based_on_probability 🔒

The code here mainly refers to the relevant implementation of opentelemetry, and determines whether the span needs to be collected based on the TraceId and sampling rate (i.e. prob).